optimize: add @pintora/diagrams/shared-grammars so others can use them
