+++ .changeset/friendly-sheep-push.md
@@ -1,0 +1,8 @@
+---
+'@pintora/diagrams': patch
+'pintora-website': patch
+'@pintora/cli': patch
+'@pintora/standalone': patch
+---
+
+optimize: add @pintora/diagrams/shared-grammars so others can use them
