+++ packages/pintora-diagrams/shared-grammars/config.d.ts
@@ -1,0 +1,17 @@
+/** action type for `@param` statement */
+export type ParamAction = {
+  type: 'addParam'
+  key: string
+  value: string
+}
+
+/** action type for `@config` statement */
+export type OverrideConfigAction<T = unknown> =
+  | {
+      type: 'overrideConfig'
+      value: T
+    }
+  | {
+      type: 'overrideConfig'
+      error: Error
+    }
