+++ packages/pintora-diagrams/shared-grammars/whitespace.ne
