+++ packages/pintora-diagrams/src/component/artist.ts
@@ -313,6 +313,7 @@ function drawGroupsTo(parentMark: Group, ir: ComponentDiagramIR, g: LayoutGraph)
               lineWidth: conf.groupBorderWidth,
             },
           })
+          console.log('bgMark', bgMark)
           if (bgMark) {
             // console.log('bgMark', groupId, bgMark, 'bounds', bgMark.symbolBounds)
             // node.outerTop = bgMark.symbolBounds.top + y
